requirements.txt
-i https://pypi.org/simple
adafruit-dht==1.4.0
rpi.gpio==0.7.0